pyproject.toml
[project]
name = "multistage-rag"
version = "1.0.0"
description = "A configurable multi-stage RAG system"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "MultiStage-RAG Team", email = "multistage-rag@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # 核心框架
    "fastapi>=0.104.1,<0.105.0",
    "uvicorn[standard]>=0.24.0,<0.25.0",
    "pydantic>=2.5.0,<3.0.0",
    "pydantic-settings>=2.1.0,<3.0.0",
    "python-multipart>=0.0.6,<0.1.0",

    # 向量存储（默认：Chroma）
    "chromadb>=0.4.22,<0.5.0",

    # 重排序（默认：阿里百炼）
    "dashscope>=1.14.0,<2.0.0",

    # 缓存（默认：Redis）
    "redis>=5.0.1,<6.0.0",

    # NLP 处理
    "sentence-transformers>=2.2.2,<3.0.0",
    "transformers>=4.36.0,<5.0.0",
    "torch>=2.1.0,<3.0.0",
    "jieba>=0.42.1,<0.43.0",

    # 数据科学
    "numpy>=1.26.2,<2.0.0",
    "scikit-learn>=1.3.2,<2.0.0",
    "pandas>=2.1.4,<3.0.0",

    # 异步和网络
    "httpx>=0.25.2,<0.26.0",
    "aiohttp>=3.9.1,<4.0.0",
    "tenacity>=8.2.3,<9.0.0",
    "anyio>=4.2.0,<5.0.0",

    # 配置和工具
    "pyyaml>=6.0.1,<7.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "psutil>=5.9.6,<6.0.0",

    # 监控和日志
    "prometheus-client>=0.19.0,<0.20.0",
    "colorlog>=6.8.0,<7.0.0",
]

[project.optional-dependencies]
# 开发依赖
dev = [
    "pytest>=7.4.3,<8.0.0",
    "pytest-asyncio>=0.21.1,<0.22.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "black>=23.12.1,<24.0.0",
    "mypy>=1.8.0,<2.0.0",
    "isort>=5.13.2,<6.0.0",
    "pre-commit>=3.6.0,<4.0.0",
]

# 可选组件
faiss = [
    "faiss-cpu>=1.7.4,<2.0.0",  # FAISS 向量存储
]

cohere = [
    "cohere>=4.34,<5.0.0",      # Cohere 重排序器
]

openai = [
    "openai>=1.6.1,<2.0.0",     # OpenAI LLM/重排序器
]

anthropic = [
    "anthropic>=0.25.2,<0.26.0", # Claude LLM
]

# GPU 支持（如果需要 GPU 加速）
gpu = [
    "torch>=2.1.0",  # 需要根据 CUDA 版本安装
]

[build-system]
requires = ["setuptools>=65.0", "setuptools_scm>=7.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["multistage_rag*"]

[tool.black]
line-length = 88
target-version = ['py311', 'py312']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.4"
addopts = "-ra -q --tb=short"
testpaths = [
    "tests",
    "src",
]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]